chore: run ng generate @angular/core:standalone
with Bootstrap the application using standalone APIs
- clean the generated content
- refactor the code into `app.config.ts` to match new angular projects
