+++ demo/src/app/marked-options-factory.ts
@@ -1,0 +1,13 @@
+import { MarkedOptions, MarkedRenderer } from 'ngx-markdown';
+import { AnchorService } from '@shared/anchor/anchor.service';
+
+export function markedOptionsFactory(anchorService: AnchorService): MarkedOptions {
+  const renderer = new MarkedRenderer();
+
+  // fix `href` for absolute link with fragments so that _copy-paste_ urls are correct
+  renderer.link = (href: string, title: string, text: string) => {
+    return MarkedRenderer.prototype.link.call(renderer, anchorService.normalizeExternalUrl(href), title, text);
+  };
+
+  return { renderer };
+}
