--- demo/src/app/shared/anchor/anchor.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { AnchorService } from './anchor.service';
-
-@NgModule({
-  providers: [AnchorService],
-})
-export class AnchorModule { }
